Made structure more elegant
