+++ Core/database.py
@@ -1,0 +1,6 @@
+from motor.motor_asyncio import AsyncIOMotorClient
+
+MONGO_URI = "mongodb://localhost:27017"
+client = AsyncIOMotorClient(MONGO_URI)
+
+db = client.mcore
